Setup FizzBuzz
At pg. 48 it says; 

“It's not doing much of anything, but what we have here is, in fact, a full-fledged test case. We could execute our test suite right now, and the framework would run it.”

It took me some back and forth to previous recipe to get my test suite to run without an error. Since these first few are targeting beginners and it’s not much code, I’d suggest showing a code summary of what is needed to type ‘rake’ and there be no issues.